Highfever/requirements.txt
pytest==7.1.2
selenium==4.4.3
requests==2.28.1
pyjwt~=2.5.0
authenticator==1.1.3
pyotp==2.7.0
pytest-html=-3.1.1
allure-==2.19.0





